Create an object AssessmentTest

Create an AssessmentTest rqti-object.

Usage

assessmentTest(
  section,
  identifier = generate_id(type = "test"),
  title = identifier,
  time_limit = 90L,
  max_attempts = 1L,
  academic_grading = FALSE,
  grade_label = c(en = "Grade", de = "Note"),
  table_label = c(en = "Grade", de = "Note"),
  navigation_mode = "nonlinear",
  submission_mode = "individual",
  allow_comment = TRUE,
  rebuild_variables = TRUE,
  metadata = qti_metadata(),
  points = NA_real_
)

Arguments

section

A list containing AssessmentSection objects.

identifier

A character value indicating the identifier of the test file. By default, it is generated as 'id_test_dddd', where dddd represents random digits.

title

A character value, optional, representing the file title. By default, it takes the value of the identifier.

time_limit

An integer value, optional, controlling the time given to a candidate for the test in minutes. Default is 90 minutes.

max_attempts

An integer value, optional, indicating the maximum number of attempts allowed for the candidate. Default is 1.

academic_grading

A boolean, optional; enables showing a grade to the candidate at the end of the testing according to the 5-point academic grade system as feedback. Default is FALSE.

grade_label

A character value, optional; a short message that shows with a grade in the final feedback; for multilingual use, it can be a named vector with two-letter ISO language codes as names (e.g., c(en="Grade", de="Note")); during test creation, it takes the value for the language of the operating system; c(en="Grade", de="Note")is default.

table_label

A character value, optional; a concise message to display as the column title of the grading table in the final feedback; for multilingual use, it can be a named vector with two-letter ISO language codes as names (e.g., c(en="Grade", de="Note")); during test creation, it takes the value for the language of the operating system; c(en="Grade", de="Note")is default.

navigation_mode

A character value, optional, determining the general paths that the candidate may have during the exam. Two mode options are possible: - 'linear': Candidate is not allowed to return to previous questions. - 'nonlinear': Candidate is free to navigate; used by default.

submission_mode

A character value, optional, determining when the candidate's responses are submitted for response processing. One of two mode options is possible: - 'individual': Submit candidates' responses on an item-by-item basis; used by default. - 'simultaneous': Candidates' responses are submitted all together by the end of the test.

allow_comment

A boolean, optional, enabling the candidate to leave comments in each question. Default is TRUE.

rebuild_variables

A boolean, optional, enabling the recalculation of variables and reshuffling the order of choices for each item-attempt. Default is TRUE.

metadata

An object of class QtiMetadata that holds metadata information about the test. By default it creates QtiMetadata object. See qti_metadata().

points

Do not use directly; the maximum number of points for the exam/test. It is calculated automatically as a sum of points of included tasks.

Value

An AssessmentTest object.

See also

Examples

sc <- sc <- singleChoice(prompt = "Question", choices = c("A", "B", "C"))
es <- new("Essay", prompt = "Question")
s <- section(c(sc, es), title = "Section with nonrandomized tasks")
t <- assessmentTest(list(s), title = "Example of the Exam")